refractor : moving fragments into the corresponding packages
