+++ ZooDeLille/app/src/main/java/com/example/zoodelille/view/info/fragment/InfoActivity.java
@@ -1,9 +1,11 @@
-package com.example.zoodelille;
+package com.example.zoodelille.view.info.fragment;
 
 import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
+
+import com.example.zoodelille.R;
 
 import androidx.appcompat.app.AppCompatActivity;
