+++ ZooDeLille/app/src/main/java/com/example/zoodelille/view/qrcode/fragment/QRCodeFragment.java
@@ -1,9 +1,11 @@
-package com.example.zoodelille;
+package com.example.zoodelille.view.qrcode.fragment;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import com.example.zoodelille.R;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
