+++ ZooDeLille/app/src/main/java/com/example/zoodelille/view/route/fragment/RouteFragment.java
@@ -1,4 +1,4 @@
-package com.example.zoodelille;
+package com.example.zoodelille.view.route.fragment;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -7,41 +7,47 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Toolbar;
 
+import com.example.zoodelille.R;
+
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
 
-public class AnimalFragment extends Fragment {
-    private static final String name = "Animal";
+public class RouteFragment extends Fragment {
+    private static final String name = "Route";
     private View m_view;
     private Toolbar toolbar;
 
-    public static AnimalFragment newInstance(){
-        return new AnimalFragment();
+    public static RouteFragment newInstance(){
+        return new RouteFragment();
     }
 
     @Nullable
     @Override
     public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
         super.onCreateView(inflater, container, savedInstanceState);
-        m_view = inflater.inflate(R.layout.fragment_animal, container, false);
+        m_view = inflater.inflate(R.layout.fragment_routes, container, false);
         return m_view;
     }
 
     @Override
     public void onActivityCreated(@Nullable Bundle savedInstanceState) {
         super.onActivityCreated(savedInstanceState);
-        toolbar = (Toolbar) m_view.findViewById(R.id.toolbar_animals);
-        toolbar.inflateMenu(R.menu.animals_filter);
-        toolbar.setTitle(R.string.animals);
+        toolbar = (Toolbar) m_view.findViewById(R.id.toolbar_routes);
+        toolbar.inflateMenu(R.menu.routes_filter);
+        toolbar.setTitle(R.string.routes);
         toolbar.setOnMenuItemClickListener(new Toolbar.OnMenuItemClickListener() {
             @Override
             public boolean onMenuItemClick(MenuItem item) {
-                if(item.getItemId()==R.id.animals_filter_alpha)
+                if(item.getItemId()==R.id.routes_filter_alpha)
                 {
                     // do something
                 }
-                else if(item.getItemId()== R.id.animals_filter_like)
+                else if(item.getItemId()== R.id.routes_filter_like)
+                {
+                    // do something
+                }
+                else if(item.getItemId()== R.id.routes_filter_distance)
                 {
                     // do something
                 }
